refactor(user-skills): extract bundle import owner
